Provide retry timeout for off signal and adjust debug log messages
If the radio was not working (or we're out of range), then the logic in
PresenseSensor::report() would cause the sensor to continuously retry to
send the off signal. The logic used would also cause ::motionDetected()
to never report false.

Now we provide a mechanism for retrying a set number of times before
moving on. ::motionDetected() will still report true for a little too
long, but it should be acceptable now.

In doing this, I also adjusted the log messages to be a little easier to
read.